requirements.txt
Django==1.7.7
coverage==3.7.1
django-nose==1.3
factory-boy==2.5.2
nose==1.3.4
pep8==1.6.2
wsgiref==0.1.2
